+++ src/app/app.module.ts
@@ -7,22 +7,15 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { MainLayoutComponent } from './layouts/main-layout/main-layout.component';
 import { AngularFireModule } from '@angular/fire/compat';
-import { AngularFireDatabaseModule } from '@angular/fire/compat/database';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ToastrModule } from 'ngx-toastr';
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { TranslateHttpLoader, provideTranslateHttpLoader } from '@ngx-translate/http-loader';
-import { ServiceWorkerModule } from '@angular/service-worker';
 import { AdminLayoutComponent } from './layouts/admin-layout/admin-layout.component';
 import { EmptyLayoutComponent } from './layouts/empty-layout/empty-layout.component';
 import {AngularFireAnalyticsModule} from '@angular/fire/compat/analytics';
 import {ReactiveFormsModule} from '@angular/forms';
-import { MessagingComponent } from './modals/messaging/messaging.component';
-import { AngularFireMessagingModule } from '@angular/fire/compat/messaging';
-import {MessagingService} from './core/services-firebase';
-import {AsyncPipe} from '@angular/common';
-import { TutorialComponent } from './modals/tutorial/tutorial.component';
 import { GeolocationComponent } from './modals/geolocation/geolocation.component';
 
 
@@ -32,8 +25,6 @@ import { GeolocationComponent } from './modals/geolocation/geolocation.component
     MainLayoutComponent,
     AdminLayoutComponent,
     EmptyLayoutComponent,
-    MessagingComponent,
-    TutorialComponent,
     GeolocationComponent,
   ],
   imports: [
@@ -42,9 +33,7 @@ import { GeolocationComponent } from './modals/geolocation/geolocation.component
     BrowserAnimationsModule,
     SharedModule,
     AngularFireModule.initializeApp(environment.firebase),
-    AngularFireDatabaseModule,
     AngularFireAuthModule,
-    AngularFireMessagingModule,
     AngularFireAnalyticsModule,
     ToastrModule.forRoot({
       timeOut: 7000,
@@ -57,8 +46,6 @@ import { GeolocationComponent } from './modals/geolocation/geolocation.component
         useClass: TranslateHttpLoader
       }
     }),
-    ServiceWorkerModule.register('ngsw-worker.js', { enabled: environment.production }),
-    // NgcCookieConsentModule.forRoot(cookieConfig),
     ReactiveFormsModule
   ],
   exports: [
@@ -67,9 +54,7 @@ import { GeolocationComponent } from './modals/geolocation/geolocation.component
   providers: [
     {provide: LOCALE_ID, useValue: 'fr' },
     provideHttpClient(withInterceptorsFromDi()),
-    provideTranslateHttpLoader({ prefix: './assets/i18n/', suffix: '.json' }),
-    MessagingService,
-    AsyncPipe
+    provideTranslateHttpLoader({ prefix: './assets/i18n/', suffix: '.json' })
   ],
   bootstrap: [AppComponent]
 })