+++ src/app/core/services-firebase/index.ts
@@ -3,4 +3,3 @@ export * from './report.service';
 export * from './base.service';
 export * from './auth.service';
 export * from './user.service';
-export * from './messaging.service';