+++ src/app/main/map/map-view/map-view.component.spec.ts
@@ -10,7 +10,6 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { RouterTestingModule } from '@angular/router/testing';
 import { HttpClientModule } from '@angular/common/http';
-import { OwlDateTimeModule, OwlNativeDateTimeModule, OWL_DATE_TIME_LOCALE } from '@danielmoncada/angular-datetime-picker';
 import { MapLegendComponent } from 'src/app/main/map/components/map-legend/map-legend.component';
 import { NgbTooltipConfig, NgbTooltip } from '@ng-bootstrap/ng-bootstrap';
 import {TranslateModule, TranslateService} from '@ngx-translate/core';
@@ -32,8 +31,6 @@ describe('MapViewComponent', () => {
         NgbTooltip,
         NoopAnimationsModule,
         RouterTestingModule,
-        OwlDateTimeModule,
-        OwlNativeDateTimeModule,
         HttpClientModule,
         FormsModule,
         ReactiveFormsModule,
@@ -48,7 +45,6 @@ describe('MapViewComponent', () => {
       providers: [
         NgbTooltipConfig,
         TranslateService,
-        {provide: OWL_DATE_TIME_LOCALE, useValue: 'fr'},
       ],
     })
     .compileComponents();