+++ src/app/main/static-page/home/home.component.ts
@@ -1,4 +1,3 @@
-import { TutorialComponent } from '../../../modals/tutorial/tutorial.component';
 import {Component, OnInit} from '@angular/core';
 import {Logger} from '@Services/logger.service';
 import {faPlayCircle, faBullhorn, faCheckCircle} from '@fortawesome/free-solid-svg-icons';
@@ -83,11 +82,6 @@ export class HomeComponent implements OnInit {
       new MetaTag(METATAG.KEYWORDS, 'lightcutoff, service information, light cut off, coupure lumiere, electricity services, service d\'electricité, no electricity, pas d\'electricité, lumiere, light, electricity, electricité, Eneo, cameroun, cameroon, energy, energie, fournisseur d’électricité, Electricité cameroun, Particuliers, entreprises, professionnels, industriels, Electricity Cameroon, ménages, actualité, Economie d\'énergie, courant, courant electrique, Logo lightcutoff, délestages, coupures, signaler coupure, signalez coupure de lumiere, rapport de coupure de lumiere, rapport, panne de courant, panne electrique, panne, report light cut off, que faire pendant une coupure de lumiere, page d\'accueil, homepage'),
       new MetaTag(METATAG.DESCRIPTION, this.translateService.instant('core.home.desc_page'))
     ]);
-  }
-
-  openModal() {
-    this.analytics.logEvent('tutorial_video');
-    this.modalService.open(TutorialComponent);
   }
 
   private loadReports() {