+++ src/app/shared/main-header/main-header.component.ts
@@ -1,10 +1,8 @@
-import { AuthService } from '../../core/services-firebase';
 import {Component, Input, OnInit} from '@angular/core';
 import { Const } from 'src/environments/const';
 import { Logger } from '@Services/logger.service';
 import {Router} from '@angular/router';
 import {environment} from '../../../environments/environment';
-import {faUserCircle} from '@fortawesome/free-solid-svg-icons';
 
 const log = new Logger('main-header.component');
 
@@ -17,25 +15,19 @@ const log = new Logger('main-header.component');
 export class MainHeaderComponent implements OnInit {
   @Input() fixedTop = false;
   @Input() isGlobalMessage = false;
-  readonly faUserCircle = faUserCircle;
   readonly isModuleAdmin = environment.app.modules.admin;
-  readonly isModuleOwnerReport = environment.app.modules.ownerReport;
-  readonly isModuleUser = environment.app.modules.user;
   readonly appTitle = Const.app.title;
 
   mapActive = false;
   adminActive = false;
-  user: any;
 
   constructor(
-    private authService: AuthService,
     private router: Router,
   ) {}
 
   ngOnInit(): void {
     log.debug('init');
     this.activeMenuDashboard();
-    this.user = this.authService.getUserLogged();
   }
 
   private activeMenuDashboard() {
@@ -43,11 +35,5 @@ export class MainHeaderComponent implements OnInit {
     const url = this.router.url;
     const route2 = url.split('/')[1];
     this.adminActive = route2 === 'admin';
-  }
-
-  logout() {
-    this.authService.logout();
-    this.user = null;
-    this.router.navigate(['/']);
   }
 }