+++ src/environments/environment.ts
@@ -3,15 +3,11 @@ export const environment = {
   environment: 'dev',
   domain: 'undefined',
   googleMapsApiKey: 'GOOGLE_MAPS_API_KEY',
-  vapidPublicKey: 'VAPID_PUBLIC_KEY',
   app: {
   modules: {
-  ownerReport: true,
   mapMenu: true,
   mapFilter: true,
   mapSearch: true,
-  user: true,
-  messaging: true,
   admin: true
   }
   },